build script antimony updated to 16.04

+++ antimony.sh
@@ -29,7 +29,16 @@ fi
 echo "---------------------------------------"
 echo "install dependencies"
 echo "---------------------------------------"
-sudo apt-get -y install libqt4-core libqt4-dev
+
+# qt
+if lsb_release -r -s | grep -q '14.04'; then
+  echo "Ubuntu 14.04"
+  sudo apt-get -y install libqt4-core libqt4-dev
+fi
+if lsb_release -r -s | grep -q '16.04'; then
+  echo "Ubuntu 16.04"
+  sudo apt-get -y install qt4-default
+fi
 
 echo "--------------------------------------"
 echo "pull repository"
@@ -58,9 +67,9 @@ mkdir $ANTIMONY_BUILD
 # CMAKE_CXX_FLAGS = -fPIC
 cd $ANTIMONY_BUILD
 # with cellml
-cmake -DWITH_CELLML=ON -DCELLML_API_INSTALL_DIR="$HOME/tmp/cellml-sdk-1.13-Linux-x86_64" -DWITH_PYTHON=ON ${SVN_DIR}/$CODE/antimony
+# cmake -DWITH_CELLML=ON -DCELLML_API_INSTALL_DIR="$HOME/tmp/cellml-sdk-1.13-Linux-x86_64" -DWITH_PYTHON=ON ${SVN_DIR}/$CODE/antimony
 # without cellml
-# cmake -DWITH_CELLML=OFF -DWITH_PYTHON=ON ${SVN_DIR}/$CODE/antimony
+cmake -DWITH_CELLML=OFF -DWITH_PYTHON=ON ${SVN_DIR}/$CODE/antimony
 make
 
 echo "--------------------------------------"
